+++ R/7.1-plCV.R
@@ -88,7 +88,7 @@ plCV <- function(array, probes, how, fold, ...){
     # Build machine and deploy
     mach <- do.call(what = how, args = args.v)
     pred <- predict(mach, array.valid, verbose = FALSE)
-    accs[v] <- calcStats(pred, array.valid, aucSkip = TRUE, plotSkip = TRUE)$acc
+    accs[v] <- calcStats(pred, aucSkip = TRUE, plotSkip = TRUE)$acc
 
     cat("plCV", v, "accuracy:", accs[v], "\n")
   }